+++ examples/remix/.wundergraph/operations/users/update.ts
@@ -1,0 +1,15 @@
+import { createOperation, z } from '../../generated/wundergraph.factory';
+
+export default createOperation.mutation({
+	input: z.object({
+		id: z.string(),
+		name: z.string(),
+		bio: z.string(),
+	}),
+	requireAuthentication: true,
+	handler: async ({ input }) => {
+		return {
+			...input,
+		};
+	},
+});