+++ packages/sdk/src/client/types.ts
@@ -150,32 +150,41 @@ export interface FetchUserRequestOptions {
 	revalidate?: boolean;
 }
 
-export interface User<Role extends string = string> {
+export interface User<Role extends string = any, CustomClaims extends {} = {}> {
 	provider?: string;
 	providerId?: string;
-	email?: string;
-	emailVerified?: boolean;
+	userId?: string;
 	name?: string;
 	firstName?: string;
 	lastName?: string;
+	middleName?: string;
 	nickName?: string;
-	description?: string;
-	userId?: string;
-	avatarUrl?: string;
+	preferredUsername?: string;
+	profile?: string;
+	picture?: string;
+	website?: string;
+	email?: string;
+	emailVerified?: boolean;
+	gender?: string;
+	birthDate?: string;
+	zoneInfo?: string;
+	locale?: string;
 	location?: string;
+
 	roles?: Role[];
 	customAttributes?: string[];
 	customClaims?: {
 		[key: string]: any;
-	};
+	} & CustomClaims;
 	accessToken?: JSONObject;
+	rawAccessToken?: string;
 	idToken?: JSONObject;
 	rawIdToken?: string;
 }
 
 export interface LogoutOptions {
 	/**
-	 * Wether to log out the user from the OpenID Connect provider.
+	 * Whether to log out the user from the OpenID Connect provider.
 	 * Some providers might require the user to visit a URL. See
 	 * the redirect field.
 	 */