+++ packages/sdk/src/codegen/templates/typescript/helpers.ts
@@ -1,6 +1,6 @@
 import type { GraphQLOperation } from '../../../graphql/operations';
 import type { ResolvedApplication } from '../../../configure';
-import { OperationType } from '@wundergraph/protobuf';
+import { OperationExecutionEngine, OperationType } from '@wundergraph/protobuf';
 
 export const isNotInternal = (op: GraphQLOperation): boolean => !op.Internal;
 
@@ -17,6 +17,12 @@ export const hasInjectedInput = (op: GraphQLOperation): boolean =>
 
 const filteredOperations = (application: ResolvedApplication, includeInternal: boolean) =>
 	includeInternal ? application.Operations : application.Operations.filter((op) => !op.Internal);
+
+export const filterNodeJSOperations = (application: ResolvedApplication): ResolvedApplication => {
+	const copy = JSON.parse(JSON.stringify(application)) as ResolvedApplication;
+	copy.Operations = copy.Operations.filter((op) => op.ExecutionEngine !== OperationExecutionEngine.ENGINE_NODEJS);
+	return copy;
+};
 
 export const operations = (application: ResolvedApplication, operationType: OperationType, includeInternal: boolean) =>
 	filteredOperations(application, includeInternal)