+++ packages/sdk/src/configure/index.ts
@@ -49,6 +49,7 @@ import {
 	PostResolveTransformationKind,
 	S3UploadProfile as _S3UploadProfile,
 	TypeConfiguration,
+	ValueType,
 	WebhookConfiguration,
 	WunderGraphConfiguration,
 } from '@wundergraph/protobuf';
@@ -112,6 +113,7 @@ export interface WunderGraphConfigApplicationConfig {
 		tokenBased?: {
 			providers: TokenAuthProvider[];
 		};
+		customClaims?: Record<string, CustomClaim>;
 	};
 	links?: LinkConfiguration;
 	security?: SecurityConfig;
@@ -125,6 +127,26 @@ export interface TokenAuthProvider {
 	jwksURL?: InputVariable;
 	userInfoEndpoint?: InputVariable;
 	userInfoCacheTtlSeconds?: number;
+}
+
+export interface CustomClaim {
+	/**
+	 * Path to the object inside the user payload to retrieve this claim
+	 */
+	jsonPath: string;
+
+	/** Value type
+	 *
+	 * @default 'string'
+	 */
+	type?: 'string' | 'int' | 'float' | 'boolean';
+
+	/** If required is true, users without this claim will
+	 * fail to authenticate
+	 *
+	 * @default true
+	 */
+	required?: boolean;
 }
 
 export interface SecurityConfig {
@@ -176,6 +198,11 @@ interface ResolvedDeployment {
 }
 
 export interface S3UploadProfile {
+	/** Whether authentication is required to upload to this profile
+	 *
+	 * @default true
+	 */
+	requireAuthentication?: boolean;
 	/** JSON schema for metadata */
 	meta?: ZodType | object;
 	/**
@@ -228,6 +255,8 @@ interface ResolvedS3UploadConfiguration extends Omit<S3UploadConfiguration, 'upl
 }
 
 export interface ResolvedWunderGraphConfig {
+	// XXX: ResolvedWunderGraphConfig is hashed by several templates.
+	// DO NOT INCLUDE UNSTABLE DATA (paths, times, etc...) in it.
 	application: ResolvedApplication;
 	deployment: ResolvedDeployment;
 	sdkVersion: string;
@@ -235,6 +264,7 @@ export interface ResolvedWunderGraphConfig {
 		roles: string[];
 		cookieBased: AuthProvider[];
 		tokenBased: TokenAuthProvider[];
+		customClaims: Record<string, CustomClaim>;
 		authorizedRedirectUris: ConfigurationVariable[];
 		authorizedRedirectUriRegexes: ConfigurationVariable[];
 		hooks: {
@@ -259,7 +289,11 @@ export interface ResolvedWunderGraphConfig {
 	serverOptions?: ResolvedServerOptions;
 }
 
-export interface CodeGenerationConfig extends ResolvedWunderGraphConfig {
+export interface CodeGenerationConfig {
+	// Keep ResolvedWunderGraphConfig in a separate field, so it can be
+	// hashed without using any unstable data as the hash input (e.g.
+	// paths like outPath or wunderGraphDir).
+	config: ResolvedWunderGraphConfig;
 	outPath: string;
 	wunderGraphDir: string;
 }
@@ -314,8 +348,16 @@ const resolveConfig = async (config: WunderGraphConfigApplicationConfig): Promis
 	}
 
 	const roles = config.authorization?.roles || ['admin', 'user'];
+	const customClaims = Object.keys(config.authentication?.customClaims ?? {});
 
-	const resolved = await resolveApplication(roles, config.apis, cors, config.s3UploadProvider || []);
+	const resolved = await resolveApplication(
+		roles,
+		customClaims,
+		config.apis,
+		cors,
+		config.s3UploadProvider || [],
+		config.server?.hooks
+	);
 
 	for (const ds of resolved.EngineConfiguration.DataSources) {
 		if (ds.Kind === DataSourceKind.GRAPHQL) {
@@ -348,6 +390,7 @@ const resolveConfig = async (config: WunderGraphConfigApplicationConfig): Promis
 			roles,
 			cookieBased: cookieBasedAuthProviders,
 			tokenBased: config.authentication?.tokenBased?.providers || [],
+			customClaims: config.authentication?.customClaims || {},
 			authorizedRedirectUris:
 				config.authentication?.cookieBased?.authorizedRedirectUris?.map((stringOrEnvironmentVariable) => {
 					if (typeof stringOrEnvironmentVariable === 'string') {
@@ -567,6 +610,7 @@ const resolveUploadConfiguration = (
 			const profileHooks = configurationHooks ? configurationHooks[key] : undefined;
 
 			uploadProfiles[key] = {
+				requireAuthentication: profile.requireAuthentication ?? true,
 				maxAllowedUploadSizeBytes: profile.maxAllowedUploadSizeBytes ?? -1,
 				maxAllowedFiles: profile.maxAllowedFiles ?? -1,
 				allowedMimeTypes: profile.allowedMimeTypes ?? [],
@@ -585,13 +629,14 @@ const resolveUploadConfiguration = (
 
 const resolveApplication = async (
 	roles: string[],
+	customClaims: string[],
 	apis: Promise<Api<any>>[],
 	cors: CorsConfiguration,
 	s3?: S3Provider,
 	hooks?: HooksConfiguration
 ): Promise<ResolvedApplication> => {
 	const resolvedApis = await Promise.all(apis);
-	const merged = mergeApis(roles, ...resolvedApis);
+	const merged = mergeApis(roles, customClaims, ...resolvedApis);
 	const s3Configurations = s3?.map((config) => resolveUploadConfiguration(config, hooks)) || [];
 	return {
 		EngineConfiguration: merged,
@@ -901,6 +946,17 @@ export const configureWunderGraphApplication = (config: WunderGraphConfigApplica
 const total = 4;
 let doneCount = 0;
 
+const mapRecordValues = <TKey extends string | number | symbol, TValue, TOutputValue>(
+	record: Record<TKey, TValue>,
+	fn: (key: TKey, value: TValue) => TOutputValue
+): Record<TKey, TOutputValue> => {
+	let output: Record<TKey, TOutputValue> = {} as any;
+	for (const key in record) {
+		output[key] = fn(key, record[key]);
+	}
+	return output;
+};
+
 const done = () => {
 	doneCount++;
 	Logger.info(`${doneCount}/${total} done`);
@@ -978,6 +1034,7 @@ const ResolvedWunderGraphConfigToJSON = (config: ResolvedWunderGraphConfig): str
 							throw new Error(`error serializing JSON schema for upload profile ${provider.name}/${key}: ${e}`);
 						}
 						uploadProfiles[key] = {
+							requireAuthentication: resolved.requireAuthentication,
 							maxAllowedUploadSizeBytes: resolved.maxAllowedUploadSizeBytes,
 							maxAllowedFiles: resolved.maxAllowedFiles,
 							allowedMimeTypes: resolved.allowedMimeTypes,
@@ -1113,10 +1170,39 @@ const resolveOperationsConfigurations = async (
 	loadedOperations: LoadOperationsOutput,
 	customJsonScalars: string[]
 ): Promise<ParsedOperations> => {
+	const customClaims = mapRecordValues(config.authentication.customClaims ?? {}, (key, claim) => {
+		let claimType: ValueType;
+		switch (claim.type) {
+			case 'string':
+				claimType = ValueType.STRING;
+				break;
+			case 'int':
+				claimType = ValueType.INT;
+				break;
+			case 'float':
+				claimType = ValueType.FLOAT;
+				break;
+			case 'boolean':
+				claimType = ValueType.FLOAT;
+				break;
+			case undefined:
+				claimType = ValueType.STRING;
+				break;
+			default:
+				throw new Error(`customClaim ${key} has invalid type ${claim.type}`);
+		}
+		return {
+			name: key,
+			jsonPathComponents: claim.jsonPath.split('.'),
+			type: claimType,
+			required: claim.required ?? true,
+		};
+	});
 	const graphQLOperations = parseGraphQLOperations(config.application.EngineConfiguration.Schema, loadedOperations, {
 		keepFromClaimVariables: false,
 		interpolateVariableDefinitionAsJSON: config.interpolateVariableDefinitionAsJSON,
 		customJsonScalars,
+		customClaims,
 	});
 	const nodeJSOperations: GraphQLOperation[] = [];
 	if (loadedOperations.typescript_operation_files)
@@ -1197,10 +1283,12 @@ const loadAndApplyNodeJsOperationOverrides = async (operation: GraphQLOperation)
 
 const applyNodeJsOperationOverrides = (
 	operation: GraphQLOperation,
-	overrides: NodeJSOperation<any, any, any, any, any>
+	overrides: NodeJSOperation<any, any, any, any, any, any, any, any, any>
 ): GraphQLOperation => {
 	if (overrides.inputSchema) {
-		operation.VariablesSchema = zodToJsonSchema(overrides.inputSchema) as any;
+		const schema = zodToJsonSchema(overrides.inputSchema) as any;
+		operation.VariablesSchema = schema;
+		operation.InternalVariablesSchema = schema;
 	}
 	if (overrides.liveQuery) {
 		operation.LiveQuery = {