+++ testapps/nextjs/.wundergraph/operations/users/get.ts
@@ -1,20 +1,14 @@
-import { createOperation, z } from '../../generated/wundergraph.factory';
+import { createOperation, z, AuthorizationError } from '../../generated/wundergraph.factory';
 
 export default createOperation.query({
 	input: z.object({
 		id: z.number(),
 	}),
 	handler: async (ctx) => {
-		const weather = await ctx.internalClient.queries.Weather({
-			input: {
-				forCity: 'Berlin',
-			},
-		});
 		return {
 			id: ctx.input.id,
 			userName: 'Jens Neuse',
 			bio: 'Founder of WunderGraph',
-			weather: weather.data?.getCityByName,
 		};
 	},
 });