+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.12.3...@wundergraph/swr@0.12.4) (2023-02-03)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.12.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.12.2...@wundergraph/swr@0.12.3) (2023-02-01)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.12.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.12.1...@wundergraph/swr@0.12.2) (2023-01-30)
 
 **Note:** Version bump only for package @wundergraph/swr